+++ internal/config/config.go
@@ -3,6 +3,7 @@ package config
 import (
 	"context"
 	"fmt"
+	"strings"
 
 	"github.com/hashicorp/actions-go-build/pkg/build"
 	"github.com/hashicorp/actions-go-build/pkg/crt"
@@ -104,6 +105,10 @@ func (c Config) init(rc crt.RepoContext, creator crt.Tool) (Config, error) {
 		return c, err
 	}
 
+	if strings.ToLower(c.Parameters.OS) == "windows" {
+		c.Product.ExecutableName = ensureExtension(c.Product.ExecutableName, ".exe")
+	}
+
 	primaryPaths := build.NewPrimaryDirs(c.Product, c.Parameters, creator)
 	verificationPaths := build.NewVerificationDirs(c.Product, c.Parameters, creator)
 
@@ -130,6 +135,13 @@ func (c Config) init(rc crt.RepoContext, creator crt.Tool) (Config, error) {
 	}
 
 	return c, nil
+}
+
+func ensureExtension(s, ext string) string {
+	if strings.HasSuffix(s, ext) {
+		return s
+	}
+	return s + ext
 }
 
 func (c Config) resolveReproducible() (string, error) {