+++ CHANGELOG.md
@@ -6,7 +6,9 @@ Instead, edit the files in dev/changes/, then run 'make docs' to update this fil
 
 ## Unreleased Changes (targeting v0.1.6)
 
+Changed:
 
+  - Windows executables automatically have a .exe extension.
 
 ## [v0.1.5](https://github.com/hashicorp/actions-go-build/releases/tag/v0.1.5) - October 12, 2022